chore: add `output` property in next config and optimize scripts
Updated the next.config.js file by adding the 'output' property to the NextConfig object to set the output to 'export'. This is necessary to implement static HTML export functionality in next.js projects. Additionally, the "prestart", "start", and "wrangler" scripts in package.json are also modified to improve the production build process. Removed the 'export' command in "prestart" task and added 'build' instead. In the "start" task, we now run the 'wrangler' script within a production environment. The 'wrangler' script is also updated to include compatibility date to ensure future proofing. These changes aim to provide a more efficient build process and improved compatibility for future updates.