chore: implement experimental webpack builder in Next.js configuration
The “experimental” field is added in the Next.js configuration file (next.config.js) to enable the experimental feature "webpackBuildWorker". This allows the application to offload some of the Webpack building process to a worker thread, which is expected to improve build performance and shorten the build time for the application.
WebWorkers enable background processing that can remove many of the restrictions placed on WebAssembly instantiation. So, this change is made considering the possible benefits of its implementation in a Next.js project.