ci: update build command in Github Actions workflow
Modified the build.yml file to change the deployment script from 'pnpm run export' to 'pnpm run build'. This change is necessary because the process of exporting an application is different from building it. The 'export' command is used for static site generation, while the 'build' command creates a production-ready version of the application.